+++ src/rules/__init__.py
@@ -1,2 +1,3 @@
-from .parser import RuleConfig, parse_rule_file
+from .parser import RuleConfig, parse_rule_file, build_inspection_plan
+from .plan import InspectionPlan
 from .engine import RuleEngine